# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/pedidos": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Lista todos os pedidos",
                "description": "Retorna uma lista de pedidos.",
                "operationId": "getPedidos",
                "responses": {
                    "200": {
                        "description": "Lista de pedidos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Cria um novo pedido",
                "description": "Cria um pedido associado a um cliente.",
                "operationId": "createPedido",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_cliente": {
                                        "description": "ID do cliente",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pedidos/{pedido}": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Obtém os detalhes de um pedido",
                "description": "Retorna os detalhes de um pedido específico.",
                "operationId": "getPedidoById",
                "parameters": [
                    {
                        "name": "pedido",
                        "in": "path",
                        "description": "ID do pedido a ser consultado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do pedido retornados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Deleta um pedido",
                "description": "Deleta um pedido específica",
                "operationId": "deletePedido",
                "parameters": [
                    {
                        "name": "pedido",
                        "in": "path",
                        "description": "ID do pedido a ser deletado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registro deletado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pedidos/{pedido}/produtos": {
            "post": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Adiciona produtos a um pedido",
                "description": "Adiciona uma lista de produtos ao pedido.",
                "operationId": "addProdutosToPedido",
                "parameters": [
                    {
                        "name": "pedido",
                        "in": "path",
                        "description": "ID do pedido",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "produtos": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id_produto": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantidade": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produtos adicionados ao pedido com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Remove um produto do pedido",
                "description": "Remove um produto específico de um pedido.",
                "operationId": "deleteProdutoFromPedido",
                "parameters": [
                    {
                        "name": "pedido",
                        "in": "path",
                        "description": "ID do pedido",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_produto": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto deletado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado no pedido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pedidos/{pedidoId}/produtos": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Lista os produtos de um pedido",
                "description": "Retorna a lista de produtos associados a um pedido específico.",
                "operationId": "getProdutosPorPedido",
                "parameters": [
                    {
                        "name": "pedidoId",
                        "in": "path",
                        "description": "ID do pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de produtos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Produto"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Produto": {
                "title": "Produto",
                "description": "Representação de um produto",
                "properties": {
                    "id": {
                        "description": "ID do produto",
                        "type": "integer"
                    },
                    "nome": {
                        "description": "Nome do produto",
                        "type": "string"
                    },
                    "descricao": {
                        "description": "Descrição do produto",
                        "type": "string"
                    },
                    "preco": {
                        "description": "Preço do produto",
                        "type": "number",
                        "format": "float"
                    },
                    "estoque": {
                        "description": "Quantidade em estoque do produto",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insira o token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}